Crate auteur

Auteur is a service that implements a control interface around a graph of live processing nodes.

Individual nodes can be scheduled to form a timeline, an example use case is creating and managing a “TV channel”:

Create a RTMP Destination node, start it

Create a Mixer, connect it to the destination, start it

Create Source nodes, connect them to the mixer, and schedule them to play one after the other

The service can of course be used for many other applications.

A companion application (auteur_controller) is provided as an example client.

The design of this project is similar to that of Brave, the implementation choices are different however.

Modules

config
Configuration options for the application

controller
Interface with remote controller

destination
A destination processing node.

mixer
A mixer processing node.

node
Implementation of a node management system

server
Implementation of the HTTP service

source
A source processing node.

utils
A set of utilities for all nodes to use

Functions

main
Application entry point