Struct auteur::mixer::Mixer

pub struct Mixer {
    id: String,
    pipeline: Pipeline,
    pipeline_manager: Option<Addr<PipelineManager>>,
    video_producer: StreamProducer,
    audio_producer: StreamProducer,
    consumer_slots: HashMap<String, ConsumerSlot>,
    audio_mixer: Element,
    video_mixer: Element,
    video_mixing_state: Arc<Mutex<VideoMixingState>>,
    audio_mixing_state: Arc<Mutex<AudioMixingState>>,
    state_machine: StateMachine,
    settings: HashMap<String, Arc<Mutex<Setting>>>,
}

The Mixer actor

Fields
id: String

Unique identifier

pipeline: Pipeline

The wrapped pipeline

pipeline_manager: Option<Addr<PipelineManager>>

A helper for managing the pipeline

video_producer: StreamProducer

Output video producer

audio_producer: StreamProducer

Output audio producer

consumer_slots: HashMap<String, ConsumerSlot>

Input connection points

audio_mixer: Element

audiomixer

video_mixer: Element

compositor

video_mixing_state: Arc<Mutex<VideoMixingState>>

Used for showing and hiding the base plate, and updating slot video controllers

audio_mixing_state: Arc<Mutex<AudioMixingState>>

Used for updating slot audio controllers

state_machine: StateMachine

Our state machine

settings: HashMap<String, Arc<Mutex<Setting>>>

Our output settings

Implementations

impl Mixer

fn create_settings() -> HashMap<String, Arc<Mutex<Setting>>>

TODO: potentially use this for inspectability?

fn setting(&self, name: &str) -> Option<MutexGuard<'_, Setting>>

fn settings(&self) -> HashMap<String, Value>

fn control_points(&self) -> HashMap<String, Vec<ControlPoint>>

pub fn new(
id: &str,
config: Option<HashMap<String, Value>>
) -> Result<Self, Error>

Create a mixer

fn parse_slot_config_key(property: &str) -> Result<(bool, &str), Error>

fn connect_slot(
pipeline: &Pipeline,
slot: &mut ConsumerSlot,
mixer_id: &str,
id: &str,
width: i32,
height: i32,
sample_rate: i32
) -> Result<(), Error>

Connect an input slot to compositor and audiomixer

fn build_base_plate(
&mut self,
width: i32,
height: i32
) -> Result<Element, Error>

Build the base plate. It may be either a live videotestsrc, or an imagefreeze’d image when a fallback image was specified

fn update_audio_mixing_state(
agg: &Aggregator,
id: &str,
pts: ClockTime,
mixing_state: &mut AudioMixingState
)

Update slot controllers

fn synchronize_slot_controllers(
agg: &Aggregator,
id: &str,
duration: ClockTime,
controllers: &mut HashMap<String, PropertyController>
) -> HashMap<String, PropertyController>

fn synchronize_mixer_controllers(
agg: &Aggregator,
base_plate_pad: &Pad,
id: &str,
duration: ClockTime,
controllers: &mut HashMap<String, SettingController>,
capsfilter: &Option<Element>
) -> HashMap<String, SettingController>

fn update_video_mixing_state(
agg: &Aggregator,
id: &str,
pts: ClockTime,
mixing_state: &mut VideoMixingState,
timeout: ClockTime
)

Show or hide our base plate, update slot controllers

fn start_pipeline(
&mut self,
ctx: &mut Context<Self>
) -> Result<StateChangeResult, Error>

Start our pipeline when cue_time is reached

fn connect(
&mut self,
link_id: &str,
video_producer: &StreamProducer,
audio_producer: &StreamProducer,
config: Option<HashMap<String, Value>>
) -> Result<(), Error>

Implement Connect command

fn disconnect(&mut self, slot_id: &str) -> Result<(), Error>

Implement Disconnect command

fn add_slot_control_point(
&mut self,
slot_id: &str,
property: &str,
point: ControlPoint
) -> Result<(), Error>

Implement AddControlPoint command for slots

fn remove_slot_control_point(
&mut self,
controller_id: &str,
slot_id: &str,
property: &str
)

Implement RemoveControlPoint command for slots

fn slot_control_points(
&self
) -> HashMap<String, HashMap<String, Vec<ControlPoint>>>

fn slot_settings(&self) -> HashMap<String, HashMap<String, Value>>

fn add_control_point(
&mut self,
property: String,
point: ControlPoint
) -> Result<(), Error>

Implement AddControlPoint command for the mixer

fn remove_control_point(&mut self, controller_id: &str, property: &str)

Implement RemoveControlPoint command for the mixer

fn stop(&mut self, ctx: &mut Context<Self>)