Struct auteur::source::Source

pub struct Source {
    id: String,
    uri: String,
    audio_producer: Option<StreamProducer>,
    video_producer: Option<StreamProducer>,
    media: Option<Media>,
    monitor_handle: Option<SpawnHandle>,
    state_machine: StateMachine,
}

The Source actor

Fields
id: String

Unique identifier

uri: String

URI the source will play

audio_producer: Option<StreamProducer>

Output audio producer

video_producer: Option<StreamProducer>

Output video producer

media: Option<Media>

GStreamer elements when prerolling or playing

monitor_handle: Option<SpawnHandle>

Statistics timer

state_machine: StateMachine

Our state machine

Implementations

impl Source

pub fn new(id: &str, uri: &str, audio: bool, video: bool) -> Self

Create a source

fn connect_pad(
id: String,
is_video: bool,
pipeline: &Pipeline,
pad: &Pad,
video_producer: &Option<StreamProducer>,
audio_producer: &Option<StreamProducer>
) -> Result<Option<Element>, Error>

Connect pads exposed by fallbacksrc to our output producers

fn preroll(
&mut self,
ctx: &mut Context<Self>
) -> Result<StateChangeResult, Error>

Preroll the pipeline ahead of time (by default 10 seconds before cue time)

fn unblock(
&mut self,
ctx: &mut Context<Self>
) -> Result<StateChangeResult, Error>

Unblock a prerolling pipeline

fn handle_stream_change(&mut self, ctx: &mut Context<Self>, starting: bool)

A new pad was added, or an existing pad EOS’d

fn monitor_switch(&mut self, ctx: &mut Context<Self>, switch: Element)

Track the status of a new fallbackswitch

fn log_source_status(&mut self)

Trace the status of the source for monitoring purposes

fn reinitialize(&mut self) -> Result<StateChangeResult, Error>

fn stop(&mut self, ctx: &mut Context<Self>)