Trait auteur::utils::Schedulable

pub trait Schedulable<A> where
    A: Actor<Context = Context<A>> + Schedulable<A> + Handler<ErrorMessage>, {
    fn state_machine(&self) -> &StateMachine;

    fn state_machine_mut(&mut self) -> &mut StateMachine;

    fn transition(
        &mut self, 
        ctx: &mut Context<A>, 
        target: State
    ) -> Result<StateChangeResult, Error>;

    fn node_id(&self) -> &str;

    fn next_time(&self) -> Option<DateTime<Utc>> { ... }

    fn start_schedule(
        &mut self, 
        ctx: &mut Context<A>, 
        cue_time: Option<DateTime<Utc>>, 
        end_time: Option<DateTime<Utc>>
    ) -> Result<(), Error> { ... }

    fn reschedule(
        &mut self, 
        ctx: &mut Context<A>, 
        cue_time: Option<DateTime<Utc>>, 
        end_time: Option<DateTime<Utc>>
    ) -> Result<(), Error> { ... }

    fn stop_schedule(&mut self, ctx: &mut Context<A>) { ... }

    fn update_state(&mut self, state: State) { ... }

    fn do_schedule(&mut self, ctx: &mut Context<A>) { ... }

    fn update_times(
        &mut self, 
        new_cue_time: &Option<DateTime<Utc>>, 
        new_end_time: &Option<DateTime<Utc>>
    ) -> Result<(), Error> { ... }
}

(Nodes)crate::node should implement this trait in order to drive their state progression.

IMPORTANT: this trait is only concerned about progressing the state along a timeline, nodes are expected to deal with non-scheduled stops themselves, by calling stop_scheduling and managing their state machine as they wish on teardown, see Destination for an advanced case.

Required methods

fn state_machine(&self) -> &StateMachine

Reference to the node state machine

fn state_machine_mut(&mut self) -> &mut StateMachine

Mutable reference to the node state machine

fn transition(
&mut self,
ctx: &mut Context<A>,
target: State
) -> Result<StateChangeResult, Error>

Progress to the target state

State progresses linearly, but it is authorized to return to the Initial state from Starting and nodes should be prepared for this possibility, for example a prerolling source should tear down its media elements before progressing through the normal order again

fn node_id(&self) -> &str

For tracing

Loading content...
Provided methods

fn next_time(&self) -> Option<DateTime<Utc>>

fn start_schedule(
&mut self,
ctx: &mut Context<A>,
cue_time: Option<DateTime<Utc>>,
end_time: Option<DateTime<Utc>>
) -> Result<(), Error>

Start scheduling state changes, valid only in Initial state

fn reschedule(
&mut self,
ctx: &mut Context<A>,
cue_time: Option<DateTime<Utc>>,
end_time: Option<DateTime<Utc>>
) -> Result<(), Error>

Reschedule state changes, fully valid in Initial and Starting states, only end_time may change in Started state, it is not permitted to reschedule in any other state

fn stop_schedule(&mut self, ctx: &mut Context<A>)

Stop the scheduling loop

fn update_state(&mut self, state: State)

fn do_schedule(&mut self, ctx: &mut Context<A>)

Internal scheduling loop

fn update_times(
&mut self,
new_cue_time: &Option<DateTime<Utc>>,
new_end_time: &Option<DateTime<Utc>>
) -> Result<(), Error>

Utility function to perform sanity checks on reschedule commands.

Transactional, either updates all times or none.

Loading content...