Struct auteur::node::NodeManager

pub struct NodeManager {
    nodes: HashMap<String, Node>,
    links: HashMap<String, Recipient<ConsumerMessage>>,
    consumers: HashMap<String, Recipient<ConsumerMessage>>,
    producers: HashMap<String, Recipient<GetProducerMessage>>,
    no_more_modes_sender: Option<Sender<()>>,
    listeners: HashMap<String, WeakRecipient<NodeStatusMessage>>,
}

NodeManager acts as a tracker of all nodes, and dispatches messages accordingly

Nodes can be producers, consumers or both. NodeManager knows how to make logical links from one to another, actual connection to producers is delegated to the consumers however, as they might want to only perform the connection once their state has progressed.

Links have an identifier (provided by the client), that identifier can be used by the client to perform disconnection, and are used by consumers that expose multiple consumer slots to identify the correct slot to disconnect (eg mixers)

Fields
nodes: HashMap<String, Node>

All nodes by id

links: HashMap<String, Recipient<ConsumerMessage>>

All links, link_id -> consumer recipient

consumers: HashMap<String, Recipient<ConsumerMessage>>

All consumers by id

producers: HashMap<String, Recipient<GetProducerMessage>>

All producers by id

no_more_modes_sender: Option<Sender<()>>

Used when the manager is “stopping”, to ensure all nodes have fully stopped before exiting. Useful for propagating EOS.

listeners: HashMap<String, WeakRecipient<NodeStatusMessage>>

Listeners for unit tests and potential user interfaces. Listener id -> recipient

Implementations

impl NodeManager

fn create_source(&mut self, id: &str, uri: &str) -> CommandResult

Create a Source and store it as a producer

fn create_destination(
&mut self,
id: &str,
family: &DestinationFamily
) -> CommandResult

Create a Destination and store it as a consumer

fn create_mixer(
&mut self,
id: &str,
config: Option<HashMap<String, Value>>
) -> CommandResult

Create a Mixer and store it as both a consumer and a producer

fn remove_node(&mut self, id: &str)

Remove a node from our collections by id

fn stop_node(&mut self, id: &str) -> CommandResult

Tell a node to stop, by id

fn disconnect_consumer(
&self,
consumer: &mut Recipient<ConsumerMessage>,
slot_id: String
)

Tell a node to disconnect one of its consumer slots

fn disconnect_consumers(
&mut self,
video_producer: Option<StreamProducer>,
audio_producer: Option<StreamProducer>
)

Look up all consumers for a given producer and tell them to disconnect the associated slots

fn send_start_command_future(
&mut self,
id: &str,
cue_time: Option<DateTime<Utc>>,
end_time: Option<DateTime<Utc>>
) -> ResponseActFuture<Self, CommandResult>

Start a Node

fn send_schedule_command_future(
&mut self,
id: &str,
cue_time: Option<DateTime<Utc>>,
end_time: Option<DateTime<Utc>>
) -> ResponseActFuture<Self, CommandResult>

Reschedule a Node

fn connect_future(
&mut self,
link_id: &str,
src: &str,
sink: &str,
config: Option<HashMap<String, Value>>
) -> ResponseActFuture<Self, CommandResult>

Connect a producer and a consumer

fn disconnect(&mut self, link_id: &str) -> CommandResult

Disconnect a consumer by id

fn get_info_future(
&mut self,
id: Option<&String>
) -> ResponseActFuture<Self, CommandResult>

Get the info either of a specific Node, or of all nodes

fn control_property_future(
&mut self,
controllee_id: String,
property: String,
control_point: ControlPoint
) -> ResponseActFuture<Self, CommandResult>

Add a control point for a property and keep track of the controllee

fn remove_control_point(
&mut self,
id: &str,
controllee_id: &str,
property: &str
) -> CommandResult

Remove a control point by id

fn add_listener(
&mut self,
id: String,
recipient: WeakRecipient<NodeStatusMessage>
) -> Result<(), Error>

fn notify_listeners(&mut self, message: NodeStatusMessage)