Module auteur::source

A source processing node.

The only supported source type is created with a URI. In the future generators could also be supported, for example to display a countdown.

The main complexity for this node is the starting feature: when the node is scheduled to play in the future, we spin it up 10 seconds before its cue time. Non-live sources are blocked by fallbacksrc, which only picks a base time once unblocked, and data coming from live sources is discarded by the StreamProducers until forward() is called on them.

Part of the extra complexity is due to the rescheduling feature: when rescheduling a prerolling source, we want to tear down the previous pipeline, but as logical connections are tracked by the StreamProducer elements, we want to keep their lifecycle tied to that of the source. This means appsinks may be removed from an old pipeline and moved to a new one, this is safe but needs to be kept in mind.

Structs

Media
The pipeline and various GStreamer elements that the source optionally wraps, their lifetime is not directly bound to that of the source itself

NewSourceBinMessage
Sent by the Source to notify itself that the urisourcebin was added in fallbacksrc

NewSwitchMessage
Sent by the Source to notify itself that a new fallbackswitch was added in fallbacksrc

Source
The Source actor

SourceStatusMessage
Sent by the Source to notify itself that the status of one of the monitored elements changed

StreamMessage
Sent by the Source to notify itself that a stream started or ended