Module auteur::utils

A set of utilities for all nodes to use

Re-exports

pub use pipeline_manager::PipelineManager;

pub use pipeline_manager::StopManagerMessage;

pub use pipeline_manager::WaitForEosMessage;

pub use property_controller::PropertyController;

pub use schedulable::Schedulable;

pub use schedulable::StateChangeResult;

pub use schedulable::StateMachine;

pub use setting_controller::Setting;

pub use setting_controller::SettingController;

pub use setting_controller::SettingSpec;

pub use stream_producer::StreamProducer;

Modules

pipeline_manager

Helper gst::Pipeline wrapper

property_controller

Schedule gst::Object property updates

schedulable

Control the state of nodes

setting_controller

Schedule “native” setting updates

stream_producer

Data interface between nodes

Structs

ErrorMessage

Sent from PipelineManager to nodes to signal an error

Functions

get_now

In normal operation, now is the actual system time

make_element

Wrapper around gst::ElementFactory::make with a better error message